+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "alephium-nft",
-  "version": "0.4.18",
+  "version": "0.4.21",
   "private": true,
   "scripts": {
     "dev": "NODE_OPTIONS='--inspect' next dev -p 3020",
