+++ pages/api/nft-listings-count.ts
@@ -1,11 +1,12 @@
 import Validate from 'next-api-validation'
+import type { NextApiRequest, NextApiResponse } from "next";
 import { NFTListing } from '../../utils/mongodb/models/nft-listing'
 import { connectToDatabase } from '../../utils/mongodb'
 
 connectToDatabase()
 
 const nftListingsCountHandler = Validate({
-  async get(req, res) {
+  async get(req: NextApiRequest, res: NextApiResponse) {
     try {
       const searchText = req.query.search as string
       const filterArgs = searchText ? { $text: { $search: searchText, $caseSensitive: false } } : {}
